web/requirements.txt
Flask==2.3.3
requests==2.31.0
rsa==4.9
pyyaml==6.0.1
openpyxl==3.1.2